Refactor to have dynamixel_calibration and feetech_calibration
